requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.2.0+cpu
torchvision==0.17.0+cpu
flask
gunicorn
opencv-python-headless
Pillow